catalog-bot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
google-generativeai==0.3.1
pydantic==2.5.2
pydantic-settings==2.1.0
requests==2.31.0
rembg==2.0.50
Pillow==10.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
cryptography==41.0.7
